// frontend/src/shared/api/generated/model/cartDetailStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CartDetailStatus = (typeof CartDetailStatus)[keyof typeof CartDetailStatus];

export const CartDetailStatus = {
  IDLE: 'IDLE',
  MOVING: 'MOVING',
  FOLLOWING: 'FOLLOWING',
  ERROR: 'ERROR',
} as const;